🎨 Palette: Refactor LoginPage to use LoginForm and add autocomplete
