+++ src/app/[lang]/(auth)/login/page.tsx
@@ -1,74 +1,26 @@
-"use client";
+import { getDictionary } from "@/lib/dictionaries";
+import { Locale } from "@/app/i18n-config";
+import { LoginForm } from "@/components/LoginForm";
+import { GoogleSignInButton } from "./GoogleSignInButton";
 
-import { signIn } from "next-auth/react";
-import { useState } from "react";
-
-export default function LoginPage() {
-    const [email, setEmail] = useState("");
-    const [password, setPassword] = useState("");
-
-    const handleGoogleLogin = () => {
-        signIn("google", { callbackUrl: "/" });
-    };
-
-    const handleCredentialsLogin = async (e: React.FormEvent) => {
-        e.preventDefault();
-        await signIn("credentials", {
-            email,
-            password,
-            callbackUrl: "/"
-        });
-    };
+export default async function LoginPage({
+    params,
+}: {
+    params: Promise<{ lang: string }>;
+}) {
+    const { lang } = await params;
+    const dict = await getDictionary(lang as Locale);
 
     return (
         <main className="min-h-screen flex items-center justify-center bg-gray-50 py-12 px-4 sm:px-6 lg:px-8">
             <div className="max-w-md w-full space-y-8 bg-white p-8 border rounded-xl shadow-sm">
                 <div>
-                    <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900">
-                        Sign in to your account
+                    <h2 className="mt-6 text-center text-3xl font-extrabold text-gray-900 mb-8">
+                        {dict.auth?.login_title || "Sign in to your account"}
                     </h2>
                 </div>
-                <form className="mt-8 space-y-6" onSubmit={handleCredentialsLogin}>
-                    <div className="rounded-md shadow-sm -space-y-px">
-                        <div>
-                            <label htmlFor="email" className="sr-only">Email address</label>
-                            <input
-                                id="email"
-                                name="email"
-                                type="email"
-                                autoComplete="email"
-                                required
-                                className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-t-md focus:outline-none focus:ring-black focus:border-black focus:z-10 sm:text-sm"
-                                placeholder="Email address"
-                                value={email}
-                                onChange={(e) => setEmail(e.target.value)}
-                            />
-                        </div>
-                        <div>
-                            <label htmlFor="password" className="sr-only">Password</label>
-                            <input
-                                id="password"
-                                name="password"
-                                type="password"
-                                autoComplete="current-password"
-                                required
-                                className="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500 text-gray-900 rounded-b-md focus:outline-none focus:ring-black focus:border-black focus:z-10 sm:text-sm"
-                                placeholder="Password"
-                                value={password}
-                                onChange={(e) => setPassword(e.target.value)}
-                            />
-                        </div>
-                    </div>
 
-                    <div>
-                        <button
-                            type="submit"
-                            className="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-black hover:bg-gray-800 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-black"
-                        >
-                            Sign in with Credentials
-                        </button>
-                    </div>
-                </form>
+                <LoginForm dict={dict.auth} />
                 
                 <div className="mt-6 relative">
                     <div className="absolute inset-0 flex items-center">
@@ -80,12 +32,7 @@ export default function LoginPage() {
                 </div>
 
                 <div className="mt-6">
-                    <button
-                        onClick={handleGoogleLogin}
-                        className="w-full flex justify-center py-2 px-4 border border-gray-300 rounded-md shadow-sm bg-white text-sm font-medium text-gray-500 hover:bg-gray-50"
-                    >
-                        Sign in with Google
-                    </button>
+                    <GoogleSignInButton />
                 </div>
             </div>
         </main>
